% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg_install.R
\name{check_pkg_install}
\alias{check_pkg_install}
\title{Check for package installation and install}
\usage{
check_pkg_install(pkg, lib = .libPaths()[[1L]], upgrade = FALSE,
  num_workers = 1L, ask = FALSE)
}
\arguments{
\item{pkg}{Package names or remote package specifications to install.
See \link[pkgdepends:remotes]{pkgdepends::remotes} for details about remote package
specifications.}

\item{lib}{Package library to install the packages to.}

\item{upgrade}{Whether to upgrade already installed packages to the
latest available version.}

\item{num_workers}{Number of worker processes to use.}

\item{ask}{Whether to ask for confirmation.}
}
\description{
Determine if the one or more packages are installed in the library path and,
if not, prompt the user to install them. If all packages are installed, the
function silently returns the package list invisibly.
}
