% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{give_feedback}
\alias{give_feedback}
\title{Give feedback after checking whether packages are loaded}
\usage{
give_feedback(is_installed)
}
\arguments{
\item{is_installed}{Vector of booleans indicating whether packages are loaded}
}
\value{
Nothing
}
\description{
Give feedback after checking whether packages are loaded
}
