% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{Installing pak}
\alias{Installing pak}
\title{All about installing pak.}
\description{
Read this if the default installation methods do not work for you or
if you want the release candidate or development version.

```{r child = "man/chunks/install.Rmd"}
```
}
