% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{lib_activate}
\alias{lib_activate}
\alias{lib_deactivate}
\title{Activate or deactivate a given library}
\usage{
lib_activate(lib, action = "replace")

lib_deactivate(lib)
}
\arguments{
\item{lib}{The library directory}

\item{action}{\code{[character(1)]} Should lib \code{"replace"}, \code{"prepend"} or \code{"append"} the existing library paths.}
}
\description{
Activate or deactivate a given library
}
