% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{lib_temporary}
\alias{lib_temporary}
\title{Create a session specific temporary library}
\usage{
lib_temporary(dir = tempdir(), pattern = "library-")
}
\arguments{
\item{dir}{Directory to create the library in}

\item{pattern}{a non-empty character vector giving the initial part
    of the name.}
}
\description{
Create a session specific temporary library
}
