% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{local_deps}
\alias{local_deps}
\alias{local_deps_tree}
\alias{local_dev_deps}
\alias{local_dev_deps_tree}
\title{Dependencies of a package tree}
\usage{
local_deps(root = ".", dependencies = NA)

local_deps_tree(root = ".", dependencies = NA)

local_dev_deps(root = ".", dependencies = TRUE)

local_dev_deps_tree(root = ".", dependencies = TRUE)
}
\arguments{
\item{root}{Path to the package tree.}

\item{dependencies}{Which dependencies to print. Defaults to the hard
dependencies for \code{local_deps()} and \code{local_deps_tree()} and the hard
dependencies plus the development dependencies for \code{local_dev_deps()}
and \code{local_dev_deps_tree()}.}
}
\value{
All of these functions return the dependencies in a data
frame (tibble). \code{local_deps_tree()} and \code{local_dev_deps_tree()} also
print the dependency tree.
}
\description{
Dependencies of a package tree
}
