% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{local_install_deps}
\alias{local_install_deps}
\title{Install the dependencies of a package tree}
\usage{
local_install_deps(
  root = ".",
  lib = NULL,
  upgrade = TRUE,
  ask = interactive(),
  dependencies = NA
)
}
\arguments{
\item{root}{Path to the package tree.}

\item{lib}{Character vector of library paths to consider when creating the
installation plan.
- The first library path is the target where packages will be installed.
- Additional library paths, if provided, are visible to the solver as
  candidates for satisfying dependency requirements. If a needed package is
  found here at an acceptable version, it won't be re-installed in
  `lib[1]`.
- Base and recommended packages in `.Library` are always considered, i.e.
  a recommended package is only duplicated in `lib[1]` if a newer version
  is required.}

\item{upgrade}{When `FALSE`, the default, pak does the minimum amount
  of work to give you the latest version(s) of `pkg`. It will only upgrade
  dependent packages if `pkg`, or one of their dependencies explicitly
  require a higher version than what you currently have. It will also
  prefer a binary package over to source package, even if the binary
  package is older.

  When `upgrade = TRUE`, pak will ensure that you have the latest
  version(s) of `pkg` and all their dependencies.}

\item{ask}{Whether to ask for confirmation when installing a different
version of a package that is already installed. Installations that only
add new packages never require confirmation.}

\item{dependencies}{What kinds of dependencies to install. Most commonly
one of the following values:
- `NA`: only required (hard) dependencies,
- `TRUE`: required dependencies plus optional and development
  dependencies,
- `FALSE`: do not install any dependencies. (You might end up with a
  non-working package, and/or the installation might fail.)
See [Package dependency types] for other possible values and more
information about package dependencies.}
}
\value{
Data frame, with information about the installed package(s).
}
\description{
Installs the hard dependencies of a package tree (or source package file),
without installing the package tree itself.
}
\details{
Note that development (and optional) dependencies, under `Suggests` in
`DESCRIPTION`, are not installed. If you want to install them as well,
use [local_install_dev_deps()].
}
\seealso{
Other local package trees:
\code{\link[=local_deps]{local_deps()}},
\code{\link[=local_deps_explain]{local_deps_explain()}},
\code{\link[=local_install]{local_install()}},
\code{\link[=local_install_dev_deps]{local_install_dev_deps()}},
\code{\link{local_package_trees}},
\code{\link[=pak]{pak()}}
}
\concept{local package trees}
