% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{local_install_dev_deps}
\alias{local_install_dev_deps}
\title{Install all dependencies of a package tree}
\usage{
local_install_dev_deps(root = ".", lib = .libPaths()[1],
  upgrade = FALSE, ask = interactive())
}
\arguments{
\item{root}{Path to the package tree.}

\item{lib}{Package library to install the packages to.}

\item{upgrade}{Whether to upgrade already installed packages to the
latest available version.}

\item{ask}{Whether to ask for confirmation.}
}
\description{
Installs all dependencies of a package tree (or source package file),
without installing the package tree itself. It installs the development
dependencies as well, specified in the \code{Suggests} field of
\code{DESCRIPTION}.
}
\seealso{
Other local packages: \code{\link{local_install_deps}},
  \code{\link{local_install}}
}
\concept{local packages}
