% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockfile.R
\name{local_install_lockfile}
\alias{local_install_lockfile}
\title{Install packages based on a lock file}
\usage{
local_install_lockfile(lockfile = "pkg.lock", lib = .libPaths()[1])
}
\arguments{
\item{lockfile}{Path to the lock file.}

\item{lib}{Library to carry out the installation on.}
}
\description{
Install a lock file that was created with \code{\link[=local_create_lockfile]{local_create_lockfile()}}.
}
\seealso{
Other lock files: 
\code{\link{local_create_lockfile}()}
}
\concept{lock files}
