% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{local_package_trees}
\alias{local_package_trees}
\title{About local package trees}
\description{
pak can install packages from local package trees. This is convenient
for package development. See the following functions:
\itemize{
\item \code{\link[=local_install]{local_install()}} installs a package from a package tree and all of its
dependencies.
\item \code{\link[=local_install_deps]{local_install_deps()}} installs all hard dependencies of a package.
\item \code{\link[=local_install_dev_deps]{local_install_dev_deps()}} installs all hard and soft dependencies
of a package. This function is intended for package development.
}
}
\details{
Note that the last two functions do not install the package in the
specified package tree itself, only its dependencies. This is convenient
if the package itself is loaded via some other means, e.g.
\code{devtools::load_all()}, for development.
}
\seealso{
Other local package trees: 
\code{\link{local_deps_explain}()},
\code{\link{local_deps}()},
\code{\link{local_install_deps}()},
\code{\link{local_install_dev_deps}()},
\code{\link{local_install}()},
\code{\link{pak}()}
}
\concept{local package trees}
