% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{local_pkg_install}
\alias{local_pkg_install}
\title{Install a local development package}
\usage{
local_pkg_install(path = ".", lib = .libPaths()[[1L]], num_workers = 1L)
}
\arguments{
\item{path}{to the local package}

\item{lib}{Package library to install the packages to.}

\item{num_workers}{Number of worker processes to use.}
}
\description{
Install a local development package
}
