% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-requirements.R
\name{local_system_requirements}
\alias{local_system_requirements}
\title{Query the system requirements for a dev package (and its dependencies)}
\usage{
local_system_requirements(
  os,
  os_release,
  root = ".",
  execute = FALSE,
  sudo = execute,
  echo = FALSE
)
}
\arguments{
\item{os, os_release}{The operating system and operating system release version, see
\url{https://github.com/rstudio/r-system-requirements#operating-systems} for the
list of supported operating systems.}

\item{root}{Path to the package tree.}

\item{execute, sudo}{If \code{execute} is \code{TRUE}, pak will execute the system
commands (if any). If \code{sudo} is \code{TRUE}, pak will prepend the commands with
\href{https://en.wikipedia.org/wiki/Sudo}{sudo}.}

\item{echo}{If \code{echo} is \code{TRUE} and \code{execute} is \code{TRUE}, echo the command output.}
}
\value{
A character vector of commands needed to install the system requirements for the package (invisibly).
}
\description{
Returns a character vector of commands to run that will install system
requirements for the queried operating system.
}
