% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockfile.R
\name{lockfile_install}
\alias{lockfile_install}
\title{Install packages based on a lock file}
\usage{
lockfile_install(lockfile = "pkg.lock", lib = .libPaths()[1], update = TRUE)
}
\arguments{
\item{lockfile}{Path to the lock file.}

\item{lib}{Library to carry out the installation on.}

\item{update}{Whether to online install the packages that
either not installed in \code{lib}, or a different version is installed
for them.}
}
\description{
Install a lock file that was created with \code{\link[=lockfile_create]{lockfile_create()}}.
}
\seealso{
Other lock files: 
\code{\link{lockfile_create}()}
}
\concept{lock files}
