% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{meta_summary}
\alias{meta_summary}
\alias{meta_list}
\alias{meta_update}
\alias{meta_clean}
\title{Metadata cache utilities}
\usage{
meta_summary()

meta_list(pkg = NULL)

meta_update()

meta_clean(force = FALSE)
}
\arguments{
\item{pkg}{Package names, if specified then only entries for \code{pkg}
are returned.}

\item{force}{If \code{FALSE}, then pak will ask for confirmation.}
}
\value{
\code{meta_summary()} returns a list with entries:
\itemize{
\item \code{cachepath}: absolute path of the metadata cache.
\item \code{current_db}: the file that contains the current metadata database.
It is currently an RDS file, but this might change in the future.
\item \code{raw_files}: the files that are the downloaded \verb{PACKAGES*} files.
\item \code{db_files}: all metadata database files.
\item \code{size}: total size of the metadata cache.
}

\code{meta_list()} returns a data frame of all available
packages in the configured repositories.

\code{meta_update()} returns nothing.

\code{meta_clean()} returns nothing
}
\description{
Various utilities to inspect, update and clean the metadata cache.
See the pkgcache package if you need for control over the metadata cache.
}
\details{
\code{meta_summary()} returns a summary of the metadata cache.

\code{meta_list()} lists all (or some) packages in the metadata
database.

\code{meta_update()} updates the metadata database. You don't
normally need to call this function manually, because all pak functions
(e.g. \code{\link[=pkg_install]{pkg_install()}}, \code{\link[=pkg_download]{pkg_download()}}, etc.) call it automatically,
to make sure that they use the latest available metadata.

\code{meta_clean()} deletes the whole metadata DB.
}
\section{Examples}{

Metadata cache summary:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{meta_summary()
#> $cachepath
#> [1] "/home/runner/.cache/R/pkgcache/_metadata"
#> 
#> $current_db
#> [1] "/home/runner/.cache/R/pkgcache/_metadata/pkgs-c22cd3b850.rds"
#> 
#> $raw_files
#> [1] "/home/runner/.cache/R/pkgcache/_metadata/BioCann-133378caf4/src/contrib/PACKAGES.gz"      
#> [2] "/home/runner/.cache/R/pkgcache/_metadata/BioCbooks-acae71babc/src/contrib/PACKAGES.gz"    
#> [3] "/home/runner/.cache/R/pkgcache/_metadata/BioCexp-9d0a1dae1d/src/contrib/PACKAGES.gz"      
#> [4] "/home/runner/.cache/R/pkgcache/_metadata/BioCsoft-41b228b7fc/src/contrib/PACKAGES.gz"     
#> [5] "/home/runner/.cache/R/pkgcache/_metadata/BioCworkflows-6b25871327/src/contrib/PACKAGES.gz"
#> [6] "/home/runner/.cache/R/pkgcache/_metadata/CRAN-075c426938/src/contrib/PACKAGES.gz"         
#> [7] "/home/runner/.cache/R/pkgcache/_metadata/RSPM-49260a4dff/src/contrib/PACKAGES.gz"         
#> 
#> $db_files
#> [1] "/home/runner/.cache/R/pkgcache/_metadata/pkgs-c22cd3b850.rds"
#> 
#> $size
#> [1] 111564643
}\if{html}{\out{</div>}}


The current metadata DB:

\if{html}{\out{<div class="sourceCode asciicast">}}\preformatted{meta_list()
}\if{html}{\out{</div>}}

Selected packages only:

\if{html}{\out{<div class="sourceCode asciicast">}}\preformatted{meta_list(pkg = c("shiny", "htmlwidgets"))
}\if{html}{\out{</div>}}


Update the metadata DB

\if{html}{\out{<div class="sourceCode asciicast">}}\preformatted{meta_update()
}\if{html}{\out{</div>}}


Delete the metadata DB

\if{html}{\out{<div class="sourceCode asciicast">}}\preformatted{meta_clean()
}\if{html}{\out{</div>}}
}

