% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{Package dependency types}
\alias{Package dependency types}
\title{Various types of R package dependencies}
\description{
Various types of R package dependencies
}
\details{
\eval{pak:::include_docs("pkgdepends", "docs/deps.rds")}
}
