% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{pak_install_extra}
\alias{pak_install_extra}
\title{Install all optional dependencies of pak}
\usage{
pak_install_extra(upgrade = FALSE)
}
\arguments{
\item{upgrade}{Whether to install or upgrade to the latest versions
of the optional packages.}
}
\description{
These packages are not required for any pak functionality. They are
recommended for some functions that return values that are best
used with these packages. E.g. many functions return data frames, which
are printed nicer is the pillar package is available.
}
\details{
Currently only one package is optional: pillar.
}
