% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pak-cleanup.R
\name{pak_private_library}
\alias{pak_private_library}
\title{The pak private library}
\description{
pak is an R package, and needs other R packages to do its job. These
dependencies should be kept separate from the user's "regular" package
libraries, to avoid the situation when pak needs a different version
of a package than the one in the regular library.
}
\details{
To accomplish this, pak keeps all of its dependencies in a separate
library. This library is usually in the user's cache directory.

pak creates and updates its private library, as needed: every time
pak cannot load a package from the private library, including the
obvious case when the user does not have a private library, pak will
create one.

You can use \code{\link[=pak_sitrep]{pak_sitrep()}} to list the location of the pak private
library, and \code{\link[=pak_cleanup]{pak_cleanup()}} to clean it up.
}
\seealso{
Other pak housekeeping: \code{\link{pak_cleanup}},
  \code{\link{pak_setup}}, \code{\link{pak_sitrep}}
}
\concept{pak housekeeping}
