% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pak-sitrep.R
\name{pak_sitrep}
\alias{pak_sitrep}
\title{pak SITuation REPort}
\usage{
pak_sitrep()
}
\description{
It prints
* pak version,
* platform the package was built on, and the current platform,
* the current library path,
* versions of dependencies,
* whether dependencies can be loaded.
}
\section{Examples}{

```{asciicast pak-sitrep}
pak_sitrep()
```
}

\seealso{
Other pak housekeeping:
\code{\link[=pak_cleanup]{pak_cleanup()}}
}
\concept{pak housekeeping}
