% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pak-update.R
\name{pak_update}
\alias{pak_update}
\title{Update pak itself}
\usage{
pak_update(force = FALSE)
}
\arguments{
\item{force}{Whether to force an update, even if no newer version is
available.}
}
\value{
Nothing.
}
\description{
Use this function to update the released or development version of pak.
}
