% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-cleanup.R
\name{pkg_cleanup}
\alias{pkg_cleanup}
\title{Clean up pkg caches and/or the pkg library}
\usage{
pkg_cleanup(package_cache = TRUE, metadata_cache = TRUE,
  pkg_lib = TRUE, force = FALSE)
}
\arguments{
\item{package_cache}{Whether to clean up the cache of package files.}

\item{metadata_cache}{Whether to clean up the cache of package meta
data.}

\item{pkg_lib}{Whethe to clean up the pkg package library.}

\item{force}{Do not ask for confirmation. Note that to use this function
in non-interactive mode, you have to specify \code{force = FALSE}.}
}
\description{
Clean up pkg caches and/or the pkg library
}
