% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_deps}
\alias{pkg_deps}
\title{Draw the dependency tree of a package}
\usage{
pkg_deps(pkg, dependencies = NULL)
}
\arguments{
\item{pkg}{Package name or remote package specification to resolve.}

\item{dependencies}{Dependency types. Possible values:
\itemize{
\item \code{TRUE}: This means all hard dependencies plus \code{Suggests} for \code{pkg}
and hard dependencies only for dependent packages.
\item \code{FALSE}: no dependencies at all, this does not make too much sense
for this function.
\item \code{NA} (any atomic type, so \code{NA_character_}, etc. as well): only hard
dependencies. See \code{\link[pkgdepends:pkg_dep_types_hard]{pkgdepends::pkg_dep_types_hard()}}.
\item If a list with two entries named \code{direct} and \code{indirect}, it is taken
as the requested dependency types, for \code{pkg} and dependent packages.
\item If a character vector, then it is taken as the dependency types
both for direct installations and dependent packages.
}}
}
\value{
A \code{tree} object from the cli package, see \code{\link[cli:tree]{cli::tree()}}.
The tree is also printed to the screen by default
}
\description{
Draw the dependency tree of a package
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pkg_deps("curl")
pkg_deps("r-lib/fs")
\dontshow{\}) # examplesIf}
}
