% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps-explain.R
\name{pkg_deps_explain}
\alias{pkg_deps_explain}
\title{Explain how a package depends on other packages}
\usage{
pkg_deps_explain(pkg, deps, upgrade = TRUE, dependencies = NA)
}
\arguments{
\item{pkg}{Package names or package references. E.g.
  - `ggplot2`: package from CRAN, Bioconductor or a CRAN-like repository
    in general,
  - `tidyverse/ggplot2`: package from GitHub,
  - `tidyverse/ggplot2@v3.4.0`: package from GitHub tag or branch,
  - `https://examples.com/.../ggplot2_3.3.6.tar.gz`: package from URL,
  - `.`: package in the current working directory.

  See "[Package sources]" for more details.}

\item{deps}{Package names of the dependencies to explain.}

\item{upgrade}{Whether to use the most recent available package
versions.}

\item{dependencies}{What kinds of dependencies to install. Most commonly
one of the following values:
- `NA`: only required (hard) dependencies,
- `TRUE`: required dependencies plus optional and development
  dependencies,
- `FALSE`: do not install any dependencies. (You might end up with a
  non-working package, and/or the installation might fail.)
See [Package dependency types] for other possible values and more
information about package dependencies.}
}
\value{
A named list with a print method. First entries are the
  function arguments: `pkg`, `deps`, `dependencies`, the last one is
  `paths` and it contains the results in a named list, the names are
  the package names in `deps`.
}
\description{
Extract dependency chains from `pkg` to `deps`.
}
\details{
This function is similar to [pkg_deps_tree()], but its output is easier
to read if you are only interested is certain packages (`deps`).
}
\section{Examples}{

How does dplyr depend on rlang?
```{asciicast pkg-deps-explain}
pkg_deps_explain("dplyr", "rlang")
```

How does the GH version of usethis depend on cli and ps?
```{asciicast pkg-deps-explain-2}
pkg_deps_explain("r-lib/usethis", c("cli", "ps"))
```
}

