% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_install}
\alias{pkg_install}
\title{Install packages}
\usage{
pkg_install(
  pkg,
  lib = NULL,
  upgrade = FALSE,
  ask = interactive(),
  dependencies = NA
)
}
\arguments{
\item{pkg}{Package names or package references. E.g.
  - `ggplot2`: package from CRAN, Bioconductor or a CRAN-like repository
    in general,
  - `tidyverse/ggplot2`: package from GitHub,
  - `tidyverse/ggplot2@v3.4.0`: package from GitHub tag or branch,
  - `https://examples.com/.../ggplot2_3.3.6.tar.gz`: package from URL,
  - `.`: package in the current working directory.

  See "[Package sources]" for more details.}

\item{lib}{Character vector of library paths to consider when creating the
installation plan.
- The first library path is the target where packages will be installed.
- Additional library paths, if provided, are visible to the solver as
  candidates for satisfying dependency requirements. If a needed package is
  found here at an acceptable version, it won't be re-installed in
  `lib[1]`.
- Base and recommended packages in `.Library` are always considered, i.e.
  a recommended package is only duplicated in `lib[1]` if a newer version
  is required.}

\item{upgrade}{When `FALSE`, the default, pak does the minimum amount
  of work to give you the latest version(s) of `pkg`. It will only upgrade
  dependent packages if `pkg`, or one of their dependencies explicitly
  require a higher version than what you currently have. It will also
  prefer a binary package over to source package, even if the binary
  package is older.

  When `upgrade = TRUE`, pak will ensure that you have the latest
  version(s) of `pkg` and all their dependencies.}

\item{ask}{Whether to ask for confirmation when installing a different
version of a package that is already installed. Installations that only
add new packages never require confirmation.}

\item{dependencies}{What kinds of dependencies to install. Most commonly
one of the following values:
- `NA`: only required (hard) dependencies,
- `TRUE`: required dependencies plus optional and development
  dependencies,
- `FALSE`: do not install any dependencies. (You might end up with a
  non-working package, and/or the installation might fail.)
See [Package dependency types] for other possible values and more
information about package dependencies.}
}
\value{
(Invisibly) A data frame with information about the installed
  package(s).
}
\description{
Install one or more packages and their dependencies into a single
package library.
}
\section{Examples}{

```{asciicast pkg-install-dplyr}
pkg_install("dplyr")
```

Upgrade dplyr and all its dependencies:
```{asciicast pkg-install-upgrade}
pkg_install("dplyr", upgrade = TRUE)
````

Install the development version of dplyr:
```{asciicast pkg-install-gh}
pkg_install("tidyverse/dplyr")
```

Switch back to the CRAN version. This will be fast because
pak will have cached the prior install.
```{asciicast pkg-install-cran}
pkg_install("dplyr")
```
}

\seealso{
[Get started with pak], [Package sources], [FAQ],
  [The dependency solver].

Other package functions:
\code{\link[=lib_status]{lib_status()}},
\code{\link[=pak]{pak()}},
\code{\link[=pkg_deps]{pkg_deps()}},
\code{\link[=pkg_deps_tree]{pkg_deps_tree()}},
\code{\link[=pkg_download]{pkg_download()}},
\code{\link[=pkg_remove]{pkg_remove()}},
\code{\link[=pkg_status]{pkg_status()}},
\code{\link[=pkg_sysreqs]{pkg_sysreqs()}}
}
\concept{package functions}
