% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_install_local}
\alias{pkg_install_local}
\title{Install a local development package}
\usage{
pkg_install_local(path, lib = .libPaths()[[1L]], num_workers = 1L)
}
\arguments{
\item{path}{to the local package}

\item{lib}{library to install packages into}

\item{num_workers}{Number of parallel workers to use}
}
\description{
Install a local development package
}
