% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_is_installed}
\alias{pkg_is_installed}
\title{Check if package is installed}
\usage{
pkg_is_installed(pkg_name)
}
\arguments{
\item{pkg_name}{Name of package to check}
}
\value{
TRUE if package is installed, FALSE otherwise
}
\description{
Check if package is installed
}
