% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_load}
\alias{pkg_load}
\title{Load packages}
\usage{
pkg_load(
  pkg,
  install = TRUE,
  lib = .libPaths()[[1L]],
  upgrade = FALSE,
  ask = interactive(),
  dependencies = NA
)
}
\arguments{
\item{pkg}{Package names or package references. E.g.
- `ggplot2`: package from CRAN, Bioconductor or a CRAN-like repository
  in general,
- `tidyverse/ggplot2`: package from GitHub,
- `tidyverse/ggplot2@v3.4.0`: package from GitHub tag or branch,
- `https://examples.com/.../ggplot2_3.3.6.tar.gz`: package from URL,
- `.`: package in the current working directory.}

\item{install}{Whether to install missing packages. If TRUE, missing packages
  will be installed with pkg_install(). If FALSE, only try to load the #'
  packages

  See "[Package sources]" for more details.}

\item{lib}{Package library to install the packages to. Note that _all_
dependent packages will be installed here, even if they are
already installed in another library. The only exceptions are base
and recommended packages installed in `.Library`. These are not
duplicated in `lib`, unless a newer version of a recommemded package
is needed.}

\item{upgrade}{When `FALSE`, the default, pak does the minimum amount
  of work to give you the latest version(s) of `pkg`. It will only upgrade
  dependent packages if `pkg`, or one of their dependencies explicitly
  require a higher version than what you currently have. It will also
  prefer a binary package over to source package, even it the binary
  package is older.

  When `upgrade = TRUE`, pak will ensure that you have the latest
  version(s) of `pkg` and all their dependencies.}

\item{ask}{Whether to ask for confirmation when installing a different
version of a package that is already installed. Installations that only
add new packages never require confirmation.}

\item{dependencies}{What kinds of dependencies to install. Most commonly
one of the following values:
- `NA`: only required (hard) dependencies,
- `TRUE`: required dependencies plus optional and development
  dependencies,
- `FALSE`: do not install any dependencies. (You might end up with a
  non-working package, and/or the installation might fail.)
See [Package dependency types] for other possible values and more
information about package dependencies.}
}
\value{
(Invisibly) A data frame with information about the installed
  package(s).
}
\description{
Load packages, if missing try to install one or more packages and their
dependencies into a single package library.
}
\section{Examples}{

```{asciicast pkg-install-dplyr}
pkg_load("dplyr")
```

Upgrade dplyr and all its dependencies:
```{asciicast pkg-install-upgrade}
pkg_load("dplyr", upgrade = TRUE)
````

Install the development version of dplyr:
```{asciicast pkg-install-gh}
pkg_load("tidyverse/dplyr")
```

Switch back to the CRAN version. This will be fast because
pak will have cached the prior install.
```{asciicast pkg-install-cran}
pkg_load("dplyr")
```
}

\seealso{
[Get started with pak], [Package sources], [FAQ],
  [The dependency solver].

Other package functions: 
\code{\link{lib_status}()},
\code{\link{pak}()},
\code{\link{pkg_deps_tree}()},
\code{\link{pkg_deps}()},
\code{\link{pkg_download}()},
\code{\link{pkg_install}()},
\code{\link{pkg_remove}()},
\code{\link{pkg_status}()},
\code{\link{pkg_sysreqs}()}
}
\concept{package functions}
