% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_remove}
\alias{pkg_remove}
\title{Remove installed packages}
\usage{
pkg_remove(pkg, lib = .libPaths()[[1L]])
}
\arguments{
\item{pkg}{A character vector of packages to remove.}

\item{lib}{library to remove packages from}
}
\description{
Remove installed packages
}
