% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{pkg_search}
\alias{pkg_search}
\title{Search CRAN packages}
\usage{
pkg_search(query, ...)
}
\arguments{
\item{query}{Search query string.}

\item{...}{
  Arguments passed on to \code{\link[pkgsearch:pkg_search]{pkgsearch::pkg_search}}
  \describe{
    \item{\code{from}}{Where to start listing the results, for pagination.}
    \item{\code{size}}{The number of results to list.}
  }}
}
\value{
A data frame, that is also a \code{pak_search_result} object
with a custom print method. To see the underlying table, you
can use \verb{[]} to drop the extra classes. See examples below.
}
\description{
Search the indexed database of current CRAN packages. It uses the
pkgsearch package. See that package for more details and also
\code{\link[pkgsearch:pkg_search]{pkgsearch::pkg_search()}} for pagination, more advanced searching,
etc.
}
\section{Examples}{

Simple search

\if{html}{\out{<div class="sourceCode asciicast">}}\preformatted{pkg_search("survival")
}\if{html}{\out{</div>}}

See the underlying data frame

\if{html}{\out{<div class="sourceCode asciicast">}}\preformatted{psro <- pkg_search("ropensci")
psro[]
}\if{html}{\out{</div>}}
}

