% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-install.R
\name{pkg_sitrep}
\alias{pkg_sitrep}
\title{pkg SITuation REPort}
\usage{
pkg_sitrep()
}
\description{
It prints
\itemize{
\item pkg version,
\item the current library path,
\item location of the private library,
\item whether the pkg private library exists,
\item whether the pkg private library is functional.
}
}
