% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_upgrade}
\alias{pkg_upgrade}
\title{Upgrade an installed package, and its dependencies}
\usage{
pkg_upgrade(pkg, lib = .libPaths()[[1L]], upgrade = TRUE, ...)
}
\arguments{
\item{pkg}{Package name to upgrade. Must be a package name, general
remote specifications are not allowed here.}

\item{lib}{Library to find the installed package in. The new version
of the package(s) will be installed into the same library.}

\item{upgrade}{Whether to upgrade the dependencies of \code{pkg} as well.}

\item{...}{Additional arguments are passed to \code{\link[=pkg_install]{pkg_install()}}.}
}
\description{
pak will try to upgrade the package from the same source as the
one used for the original installation E.g. if you installed the package
from a branch of a GitHub repository, then it will try to upgrade from
the same branch.
}
\seealso{
Other package functions: 
\code{\link{lib_status}()},
\code{\link{pak_package_sources}},
\code{\link{pak}()},
\code{\link{pkg_deps_tree}()},
\code{\link{pkg_deps}()},
\code{\link{pkg_download}()},
\code{\link{pkg_install}()},
\code{\link{pkg_remove}()},
\code{\link{pkg_status}()}
}
\concept{package functions}
