% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgman-cleanup.R
\name{pkgman_cleanup}
\alias{pkgman_cleanup}
\title{Clean up pkgman caches and/or the pkgman library}
\usage{
pkgman_cleanup(package_cache = TRUE, metadata_cache = TRUE,
  pkgman_lib = TRUE, force = FALSE)
}
\arguments{
\item{package_cache}{Whether to clean up the cache of package files.}

\item{metadata_cache}{Whether to clean up the cache of package meta
data.}

\item{pkgman_lib}{Whethe to clean up the pkgman package library.}

\item{force}{Do not ask for confirmation. Note that to use this function
in non-interactive mode, you have to specify \code{force = FALSE}.}
}
\description{
Clean up pkgman caches and/or the pkgman library
}
