% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgman-install.R
\name{pkgman_install_deps}
\alias{pkgman_install_deps}
\title{Install pkgman's dependencies into its private library}
\usage{
pkgman_install_deps(mode = c("auto", "download", "copy"),
  quiet = FALSE)
}
\arguments{
\item{mode}{Where to get the packages from. "download" will try to
download them from CRAN. "copy" will try to copy them from your
current "regular" package library. "auto" will try to copy first,
and if that fails, then it tries to download.}

\item{quiet}{Whether to omit messages.}
}
\value{
The path to the private library, invisibly.
}
\description{
To avoid interference between your regular R packages and pkgman's
dependencies, pkgman works off a private library, which can be created
by \code{pkgman_install_deps()}.
}
\seealso{
\code{\link[=pkgman_sitrep]{pkgman_sitrep()}}.
}
