% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgman-install.R
\name{pkgman_sitrep}
\alias{pkgman_sitrep}
\title{pkgman SITuation REPort}
\usage{
pkgman_sitrep()
}
\description{
It prints
\itemize{
\item pkgman version,
\item the current library path,
\item location of the private library,
\item whether the pkgman private library exists,
\item whether the pkgman private library is functional.
}
}
