% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{ppm_has_binaries}
\alias{ppm_has_binaries}
\title{Does PPM build binary packages for the current platform?}
\usage{
ppm_has_binaries()
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Does PPM build binary packages for the current platform?
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
system_r_platform()
ppm_has_binaries()
\dontshow{\}) # examplesIf}
}
\seealso{
The 'pkgcache and Posit Package Manager on Linux'
article at \url{https://r-lib.github.io/pkgcache/}.

Other PPM functions:
\code{\link[=ppm_platforms]{ppm_platforms()}},
\code{\link[=ppm_r_versions]{ppm_r_versions()}},
\code{\link[=ppm_repo_url]{ppm_repo_url()}},
\code{\link[=ppm_snapshots]{ppm_snapshots()}}
}
\concept{PPM functions}
