% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{ppm_snapshots}
\alias{ppm_snapshots}
\title{List all available Posit Package Manager (PPM) snapshots}
\usage{
ppm_snapshots()
}
\value{
Data frame with two columns:
- `date`: the time the snapshot was taken, a `POSIXct` vector,
- `id`:  integer id of the snapshot, this can be used in the repository
  URL.
}
\description{
List all available Posit Package Manager (PPM) snapshots
}
\details{
The repository URL of a snapshot has the following form on Windows:
```
{base}/{repo}/{id}
```
where `{base}` is the base URL for PPM (see [ppm_repo_url()]) and
`{id}` is either the date or id of the snapshot, or `latest` for
the latest snapshot. E.g. these are equivalent:
```
https://packagemanager.posit.co/cran/5
https://packagemanager.posit.co/cran/2017-10-10
```

On a Linux distribution that has PPM support, the repository URL that
contains the binary packages looks like this:
```
{base}/{repo}/__linux__/{binary_url}/{id}
```
where `{id}` is as before, and `{binary_url}` is a code name for a release
of a supported Linux distribution. See the `binary_url` column of the
result of [ppm_platforms()] for these code names.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ppm_snapshots()
\dontshow{\}) # examplesIf}
}
\seealso{
The 'pkgcache and Posit Package Manager on Linux'
  article at <`r pkgcache:::pkgdown_url()`>.

Other PPM functions: 
\code{\link{ppm_has_binaries}()},
\code{\link{ppm_platforms}()},
\code{\link{ppm_r_versions}()},
\code{\link{ppm_repo_url}()}
}
\concept{PPM functions}
