% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_create}
\alias{proj_create}
\title{Create a project that has a private package library}
\usage{
proj_create(project_root = ".")
}
\arguments{
\item{project_root}{Root directory of the project. If it does not
exist, it will be created.}
}
\description{
Create a project that has a private package library
}
\seealso{
Other project functions: \code{\link{proj_install_dev}},
  \code{\link{proj_install}}, \code{\link{proj_remove}},
  \code{\link{proj_status}}
}
\concept{project functions}
\keyword{internal}
