% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_install_dev}
\alias{proj_install_dev}
\title{Install project dependencies, including development dependencies, into
private project library}
\usage{
proj_install_dev(root = ".", upgrade = FALSE, ask = interactive())
}
\arguments{
\item{root}{Path to the project directory. A subdirectory can be given
as well. Defaults to the current working directory.}

\item{upgrade}{Whether to try to install the latest available versions
of the specified package(s) and/or all dependencies.}

\item{ask}{Whether to ask the user for confirmation. For non-interactive
installs supply\code{FALSE} here.}
}
\value{
Data frame containing data about the installed / updated
packages.
}
\description{
Install project dependencies, including development dependencies, into
private project library
}
\seealso{
Other project functions: \code{\link{proj_create}},
  \code{\link{proj_install}}, \code{\link{proj_remove}},
  \code{\link{proj_status}}
}
\concept{project functions}
\keyword{internal}
