% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_load}
\alias{proj_load}
\title{Load a project}
\usage{
proj_load(root = ".", ...)
}
\arguments{
\item{root}{Project root.}

\item{...}{Additional arguments to pass to \code{\link[pkgload:load_all]{pkgload::load_all()}}.}
}
\description{
It uses \code{\link[pkgload:load_all]{pkgload::load_all()}} to load the project, after making sure that
all dependencies of the project are loaded from the private library.
}
\keyword{internal}
