% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_remove}
\alias{proj_remove}
\title{Remove package(s) from a project}
\usage{
proj_remove(pkg, root = ".", ask = interactive())
}
\arguments{
\item{pkg}{Package(s) to remove. These can be package names or general
remote references, e.g. \code{github::r-lib/pak}. The packages are also
removed from \code{DESCRIPTION}.}

\item{root}{Path to the project directory. A subdirectory can be given
as well. Defaults to the current working directory.}

\item{ask}{Whether to ask the user for confirmation. For non-interactive
installs supply\code{FALSE} here.}
}
\description{
Removes it both from the project library and the \code{DESCRIPTION} file.
Note that it does not remove dependencies that are not needed any more.
}
\seealso{
Other project functions: \code{\link{proj_create}},
  \code{\link{proj_install_dev}},
  \code{\link{proj_install}}, \code{\link{proj_status}}
}
\concept{project functions}
\keyword{internal}
