% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_status}
\alias{proj_status}
\title{Status of packages in the project library}
\usage{
proj_status(root = ".")
}
\arguments{
\item{root}{Path to the project directory. A subdirectory can be given
as well. Defaults to the current working directory.}
}
\value{
Data frame (tibble) the contains data about the packages
installed in the project library.
}
\description{
Status of packages in the project library
}
\seealso{
Other project functions: \code{\link{proj_create}},
  \code{\link{proj_install_dev}},
  \code{\link{proj_install}}, \code{\link{proj_remove}}
}
\concept{project functions}
\keyword{internal}
