% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{Authenticated repositories}
\alias{Authenticated repositories}
\title{Authenticated repositories}
\description{
pak supports HTTP basic authentication when interacting with CRAN-like
repositories.

```{r child = "man/chunks/auth.Rmd"}
```
}
\seealso{
[repo_auth()], [ppm_sso_login()].

Other authenticated repositories:
\code{\link[=repo_auth]{repo_auth()}},
\code{\link[=repo_auth_key_get]{repo_auth_key_get()}}
}
\concept{authenticated repositories}
