% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_activate}
\alias{repo_activate}
\title{Activiate a repository}
\usage{
repo_activate(repo)
}
\arguments{
\item{...}{One or more repositories to set. If a repository is unnamed it is
assumed to be the CRAN repository.}
}
\description{
Activiate a repository
}
