% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_deactivate}
\alias{repo_deactivate}
\title{Deactivate a repository}
\usage{
repo_deactivate(repo)
}
\arguments{
\item{repo}{name of repository to deactivate. If no repositories are left
after deactivation \code{"https://cloud.r-project.org"} is set.}
}
\description{
Deactivate a repository
}
