% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_status}
\alias{repo_status}
\title{Repository status}
\usage{
repo_status()
}
\value{
A data.frame with the following columns
\itemize{
\item type - The repository type
\item url - The repository URL
\item time - Total time taken for a HEAD request to the repositiory
\item last_modified - Last time the repositiory was updated
}
}
\description{
Repository status
}
