% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysreqsdocs.R
\name{sysreqs_db_update}
\alias{sysreqs_db_update}
\title{Update the cached copy of the system requirements database}
\usage{
sysreqs_db_update()
}
\description{
Update the cached copy of the system requirements database
}
\details{
If the the cached copy is recent, then no update is attempted. See the
\code{metadata_update_after} \eval{man_config_link("configuration option")}.
}
\seealso{
Other system requirements functions: 
\code{\link{sysreqs_check_installed}()},
\code{\link{sysreqs_db_list}()},
\code{\link{sysreqs_db_match}()},
\code{\link{sysreqs_is_supported}()},
\code{\link{sysreqs_list_system_packages}()},
\code{\link{sysreqs_platforms}()}
}
\concept{system requirements functions}
