% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{config}
\alias{config}
\title{pak configuration}
\description{
Environment variables and options that modify the default pak behavior.
\subsection{Environment variables}{
\subsection{\code{PKG_BUILD_VIGNETTES}}{

TODO
}

\subsection{\code{PKG_CACHE_DIR}}{

TODO
}

\subsection{\code{PKG_CRAN_MIRROR}}{

TODO
}

\subsection{\code{PKG_DEPENDENCIES}}{

TODO
}

\subsection{\code{PKG_LIBRARY}}{

TODO
}

\subsection{\code{PKG_METADATA_CACHE_DIR}}{

TODO
}

\subsection{\code{PKG_METADATA_UPDATE_AFTER}}{

TODO
}

\subsection{\code{PKG_PLATFORMS}}{

TODO
}

\subsection{\code{PKG_R_VERSIONS}}{

TODO
}

\subsection{\code{PKG_SYSREQS}}{

TODO
}

\subsection{\code{PKG_SYSREQS_RSPM_REPO_ID}}{

TODO
}

\subsection{\code{PKG_SYSREQS_RSPM_URL}}{

TODO
}

\subsection{\code{PKG_SYSREQS_SUDO}}{

TODO
}

\subsection{\code{PKG_SYSREQS_VERBOSE}}{

TODO
}

\subsection{\code{PKG_WINDOWS_ARCHS}}{

TODO
}

}

\subsection{Options}{
\subsection{\code{Ncpus}}{

Set to the desired number of worker processes for package installation.
If not set, then pak will use the number of logical processors in the
machine.
}

\subsection{\code{pak.no_extra_messages}}{

Set to \code{TRUE} to omit pak's warnings about missing extra packages.
}

\subsection{\code{pkg.build_vignettes}}{

TODO
}

\subsection{\code{pkg.cache_dir}}{

TODO
}

\subsection{\code{pkg.cran_mirror}}{

TODO
}

\subsection{\code{pkg.dependencies}}{

TODO
}

\subsection{\code{pkg.library}}{

TODO
}

\subsection{\code{pkg.metadata_cache_dir}}{

TODO
}

\subsection{\code{pkg.metadata_update_after}}{

TODO
}

\subsection{\code{pkg.package_cache_dir}}{

TODO
}

\subsection{\code{pkg.platforms}}{

TODO
}

\subsection{\code{pkg.r_versions}}{

TODO
}

\subsection{\code{pkg.sysreqs}}{

TODO
}

\subsection{\code{pkg.sysreqs_rspm_repo_id}}{

TODO
}

\subsection{\code{pkg.sysreqs_rspm_url}}{

TODO
}

\subsection{\code{pkg.sysreqs_sudo}}{

TODO
}

\subsection{\code{pkg.sysreqs_verbose}}{

TODO
}

\subsection{\code{pkg.windows_archs}}{

TODO
}

\subsection{\code{repos}}{

The CRAN-like repositories to use. See \code{?options} for details.
}

}
}
