% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{faq}
\alias{faq}
\title{Frequently Asked Questions}
\description{
Please look at this before asking questions.
}
\section{Package installation}{
\subsection{How do I reinstall a package?}{

pak does not reinstall a package, if the same version is already installed.
Sometimes you still want a reinstall, e.g. to fix a broken installation.
In this case you can delete the package and then install it, or use the
\code{?reinstall} parameter:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{pak::pkg_install("tibble")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #172431;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#>                                                                                 
#> â†’ Will <span style="font-style: italic;">install</span> 13 packages.                                                     
#> â†’ All 13 packages (7.45 MB) are cached.                                         
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">cli</span>         3.3.0                                                             
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">crayon</span>      1.5.1                                                             
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">ellipsis</span>    0.3.2                                                             
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">fansi</span>       1.0.3                                                             
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">glue</span>        1.6.2                                                             
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">lifecycle</span>   1.0.1                                                             
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">magrittr</span>    2.0.3                                                             
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">pillar</span>      1.7.0                                                             
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">pkgconfig</span>   2.0.3                                                             
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">rlang</span>       1.0.2                                                             
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">tibble</span>      3.1.7                                                             
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">utf8</span>        1.2.2                                                             
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">vctrs</span>       0.4.1                                                             
#> <span style="color: #2AA198;">â„¹</span> No downloads are needed, 13 pkgs (7.45 MB) are cached                         
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">ellipsis</span> 0.3.2 (aarch64-apple-darwin20) (38.65 kB)                        
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">lifecycle</span> 1.0.1 (aarch64-apple-darwin20) (107.55 kB)                      
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">crayon</span> 1.5.1 (aarch64-apple-darwin20) (158.62 kB)                         
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">glue</span> 1.6.2 (aarch64-apple-darwin20) (154.64 kB)                           
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">pkgconfig</span> 2.0.3 (aarch64-apple-darwin20) (17.70 kB)                       
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">utf8</span> 1.2.2 (aarch64-apple-darwin20) (209.24 kB)                           
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">magrittr</span> 2.0.3 (aarch64-apple-darwin20) (231.25 kB)                       
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">fansi</span> 1.0.3 (aarch64-apple-darwin20) (379.23 kB)                          
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">cli</span> 3.3.0 (aarch64-apple-darwin20) (1.17 MB)                              
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">pillar</span> 1.7.0 (aarch64-apple-darwin20) (744.33 kB)                         
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">tibble</span> 3.1.7 (aarch64-apple-darwin20) (723.61 kB)                         
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">vctrs</span> 0.4.1 (aarch64-apple-darwin20) (1.76 MB)                            
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">rlang</span> 1.0.2 (aarch64-apple-darwin20) (1.79 MB)                            
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">cli</span> 3.3.0  <span style="color: #a3a3a3;">(66ms)</span>                                                   
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">crayon</span> 1.5.1  <span style="color: #a3a3a3;">(75ms)</span>                                                
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">ellipsis</span> 0.3.2  <span style="color: #a3a3a3;">(85ms)</span>                                              
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">fansi</span> 1.0.3  <span style="color: #a3a3a3;">(109ms)</span>                                                
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">glue</span> 1.6.2  <span style="color: #a3a3a3;">(117ms)</span>                                                 
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">lifecycle</span> 1.0.1  <span style="color: #a3a3a3;">(126ms)</span>                                            
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">magrittr</span> 2.0.3  <span style="color: #a3a3a3;">(133ms)</span>                                             
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">pillar</span> 1.7.0  <span style="color: #a3a3a3;">(151ms)</span>                                               
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">pkgconfig</span> 2.0.3  <span style="color: #a3a3a3;">(36ms)</span>                                             
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">rlang</span> 1.0.2  <span style="color: #a3a3a3;">(56ms)</span>                                                 
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">tibble</span> 3.1.7  <span style="color: #a3a3a3;">(40ms)</span>                                                
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">utf8</span> 1.2.2  <span style="color: #a3a3a3;">(38ms)</span>                                                  
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">vctrs</span> 0.4.1  <span style="color: #a3a3a3;">(28ms)</span>                                                 
#> <span style="color: #859900;">âœ”</span> 1 pkg + 12 deps: added 13, dld 13 (7.48 MB) <span style="color: #b8b8b8;">[4.7s]</span>                            
</pre></div>
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{pak::pkg_install("tibble?reinstall")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #172431;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#>                                                                                 
#> â†’ Will <span style="font-style: italic;">install</span> 1 package.                                                       
#> â†’ The package (724.32 kB) is cached.                                            
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">tibble</span>   3.1.7                                                                
#> <span style="color: #2AA198;">â„¹</span> No downloads are needed, 1 pkg (724.32 kB) is cached                          
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">tibble</span> 3.1.7  <span style="color: #a3a3a3;">(27ms)</span>                                                
#> <span style="color: #859900;">âœ”</span> 1 pkg + 12 deps: kept 11, added 1 <span style="color: #b8b8b8;">[746ms]</span>                                     
</pre></div>
}}

}

\subsection{How do I install a package from source?}{

To force the installation of a source package (instead of a binary
package), use the \code{?source} parameter:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{pak::pkg_install("tibble?source")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #172431;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#>                                                                                 
#> â†’ Will <span style="font-style: italic;">install</span> 1 package.                                                       
#> â†’ The package (672.34 kB) is cached.                                            
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">tibble</span>   3.1.7 ðŸ‘·ðŸ¼ðŸ”§                                                            
#> <span style="color: #2AA198;">â„¹</span> No downloads are needed, 1 pkg (672.34 kB) is cached                          
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">tibble</span> 3.1.7 (source) (674.87 kB)                                         
#> <span style="color: #2AA198;">â„¹</span> Building <span style="color: #268BD2;">tibble</span> 3.1.7                                                         
#> <span style="color: #859900;">âœ”</span> Built <span style="color: #268BD2;">tibble</span> 3.1.7 <span style="color: #a3a3a3;">(2.8s)</span>                                                     
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">tibble</span> 3.1.7  <span style="color: #a3a3a3;">(26ms)</span>                                                
#> <span style="color: #859900;">âœ”</span> 1 pkg + 12 deps: kept 11, added 1, dld 1 (674.87 kB) <span style="color: #b8b8b8;">[8.2s]</span>                   
</pre></div>
}}

}

\subsection{How do I ignore an optional dependency?}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{pak::pkg_install(
  c("tibble", "DiagrammeR=?ignore", "formattable=?ignore"),
  dependencies = TRUE
)
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #172431;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #B58900;">!</span> Unknown package parameter: <span style="color: #268BD2;">""</span> in <span style="color: #268BD2;">"formattable?=ignore"</span>.                       
#>                                                                                 
#> â†’ Will <span style="font-style: italic;">install</span> 57 packages.                                                     
#> â†’ All 57 packages (60.39 MB) are cached.                                        
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">askpass</span>       1.1                                                             
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">base64enc</span>     0.1-3                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">brio</span>          1.1.3                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">bslib</span>         0.3.1                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">cachem</span>        1.0.6                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">callr</span>         3.7.0                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">commonmark</span>    1.8.0                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">covr</span>          3.5.1                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">crosstalk</span>     1.2.0                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">curl</span>          4.3.2                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">desc</span>          1.4.1                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">diffobj</span>       0.3.5                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">digest</span>        0.6.29                                                          
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">DT</span>            0.23                                                            
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">evaluate</span>      0.15                                                            
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">fastmap</span>       1.1.0                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">fontawesome</span>   0.2.2                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">formattable</span>   0.2.1                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">fs</span>            1.5.2                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">highr</span>         0.9                                                             
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">htmltools</span>     0.5.2                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">htmlwidgets</span>   1.5.4                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">httpuv</span>        1.6.5                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">httr</span>          1.4.3                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">jquerylib</span>     0.1.4                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">jsonlite</span>      1.8.0                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">knitr</span>         1.39                                                            
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">later</span>         1.3.0                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">lazyeval</span>      0.2.2                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">mime</span>          0.12                                                            
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">openssl</span>       2.0.2                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">pkgload</span>       1.2.4                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">praise</span>        1.0.0                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">processx</span>      3.6.1                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">promises</span>      1.2.0.1                                                         
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">ps</span>            1.7.1                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">R6</span>            2.5.1                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">rappdirs</span>      0.3.3                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">Rcpp</span>          1.0.8.3                                                         
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">rematch2</span>      2.1.2                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">rex</span>           1.2.1                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">rmarkdown</span>     2.14                                                            
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">rprojroot</span>     2.0.3                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">rstudioapi</span>    0.13                                                            
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">sass</span>          0.4.1                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">shiny</span>         1.7.1                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">sourcetools</span>   0.1.7                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">stringi</span>       1.7.6                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">stringr</span>       1.4.0                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">sys</span>           3.4                                                             
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">testthat</span>      3.1.4                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">tinytex</span>       0.40                                                            
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">waldo</span>         0.4.0                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">withr</span>         2.5.0                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">xfun</span>          0.31                                                            
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">xtable</span>        1.8-4                                                           
#> <span style="color: #002B36;">+ </span><span style="color: #268BD2;">yaml</span>          2.3.5                                                           
#> <span style="color: #2AA198;">â„¹</span> No downloads are needed, 57 pkgs (60.39 MB) are cached                        
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">base64enc</span> 0.1-3 (aarch64-apple-darwin20) (33.62 kB)                       
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">askpass</span> 1.1 (aarch64-apple-darwin20) (22.91 kB)                           
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">R6</span> 2.5.1 (aarch64-apple-darwin20) (82.52 kB)                              
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">brio</span> 1.1.3 (aarch64-apple-darwin20) (43.84 kB)                            
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">praise</span> 1.0.0 (aarch64-apple-darwin20) (16.08 kB)                          
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">lazyeval</span> 0.2.2 (aarch64-apple-darwin20) (159.69 kB)                       
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">rappdirs</span> 0.3.3 (aarch64-apple-darwin20) (46.80 kB)                        
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">digest</span> 0.6.29 (aarch64-apple-darwin20) (272.50 kB)                        
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">rprojroot</span> 2.0.3 (aarch64-apple-darwin20) (100.20 kB)                      
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">rex</span> 1.2.1 (aarch64-apple-darwin20) (122.38 kB)                            
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">cachem</span> 1.0.6 (aarch64-apple-darwin20) (67.65 kB)                          
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">htmlwidgets</span> 1.5.4 (aarch64-apple-darwin20) (898.05 kB)                    
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">diffobj</span> 0.3.5 (aarch64-apple-darwin20) (1.01 MB)                          
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">crosstalk</span> 1.2.0 (aarch64-apple-darwin20) (404.80 kB)                      
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">testthat</span> 3.1.4 (aarch64-apple-darwin20) (2.89 MB)                         
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">sourcetools</span> 0.1.7 (aarch64-apple-darwin20) (135.84 kB)                    
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">fastmap</span> 1.1.0 (aarch64-apple-darwin20) (178.21 kB)                        
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">promises</span> 1.2.0.1 (aarch64-apple-darwin20) (1.80 MB)                       
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">callr</span> 3.7.0 (aarch64-apple-darwin20) (435.22 kB)                          
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">xtable</span> 1.8-4 (aarch64-apple-darwin20) (701.09 kB)                         
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">desc</span> 1.4.1 (aarch64-apple-darwin20) (327.81 kB)                           
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">sys</span> 3.4 (aarch64-apple-darwin20) (50.44 kB)                               
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">pkgload</span> 1.2.4 (aarch64-apple-darwin20) (152.65 kB)                        
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">ps</span> 1.7.1 (aarch64-apple-darwin20) (308.85 kB)                             
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">waldo</span> 0.4.0 (aarch64-apple-darwin20) (99.88 kB)                           
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">fs</span> 1.5.2 (aarch64-apple-darwin20) (589.90 kB)                             
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">jquerylib</span> 0.1.4 (aarch64-apple-darwin20) (526.05 kB)                      
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">yaml</span> 2.3.5 (aarch64-apple-darwin20) (216.59 kB)                           
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">formattable</span> 0.2.1 (aarch64-apple-darwin20) (168.24 kB)                    
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">withr</span> 2.5.0 (aarch64-apple-darwin20) (229.48 kB)                          
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">knitr</span> 1.39 (aarch64-apple-darwin20) (1.44 MB)                             
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">rematch2</span> 2.1.2 (aarch64-apple-darwin20) (43.89 kB)                        
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">commonmark</span> 1.8.0 (aarch64-apple-darwin20) (319.26 kB)                     
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">jsonlite</span> 1.8.0 (aarch64-apple-darwin20) (1.12 MB)                         
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">openssl</span> 2.0.2 (aarch64-apple-darwin20) (2.48 MB)                          
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">tinytex</span> 0.40 (aarch64-apple-darwin20) (132.40 kB)                         
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">mime</span> 0.12 (aarch64-apple-darwin20) (36.16 kB)                             
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">curl</span> 4.3.2 (aarch64-apple-darwin20) (879.88 kB)                           
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">highr</span> 0.9 (aarch64-apple-darwin20) (38.77 kB)                             
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">later</span> 1.3.0 (aarch64-apple-darwin20) (622.13 kB)                          
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">htmltools</span> 0.5.2 (aarch64-apple-darwin20) (334.97 kB)                      
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">httpuv</span> 1.6.5 (aarch64-apple-darwin20) (2.72 MB)                           
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">Rcpp</span> 1.0.8.3 (aarch64-apple-darwin20) (3.34 MB)                           
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">stringr</span> 1.4.0 (aarch64-apple-darwin20) (211.53 kB)                        
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">bslib</span> 0.3.1 (aarch64-apple-darwin20) (4.81 MB)                            
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">covr</span> 3.5.1 (aarch64-apple-darwin20) (290.71 kB)                           
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">httr</span> 1.4.3 (aarch64-apple-darwin20) (505.44 kB)                           
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">xfun</span> 0.31 (aarch64-apple-darwin20) (394.01 kB)                            
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">rstudioapi</span> 0.13 (aarch64-apple-darwin20) (274.23 kB)                      
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">rmarkdown</span> 2.14 (aarch64-apple-darwin20) (3.66 MB)                         
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">shiny</span> 1.7.1 (aarch64-apple-darwin20) (4.17 MB)                            
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">DT</span> 0.23 (aarch64-apple-darwin20) (1.58 MB)                                
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">fontawesome</span> 0.2.2 (aarch64-apple-darwin20) (1.53 MB)                      
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">evaluate</span> 0.15 (aarch64-apple-darwin20) (77.09 kB)                         
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">processx</span> 3.6.1 (aarch64-apple-darwin20) (287.81 kB)                       
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">sass</span> 0.4.1 (aarch64-apple-darwin20) (2.41 MB)                             
#> <span style="color: #859900;">âœ”</span> Got <span style="color: #268BD2;">stringi</span> 1.7.6 (aarch64-apple-darwin20) (14.47 MB)                         
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">R6</span> 2.5.1  <span style="color: #a3a3a3;">(58ms)</span>                                                    
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">askpass</span> 1.1  <span style="color: #a3a3a3;">(92ms)</span>                                                 
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">base64enc</span> 0.1-3  <span style="color: #a3a3a3;">(103ms)</span>                                            
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">brio</span> 1.1.3  <span style="color: #a3a3a3;">(113ms)</span>                                                 
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">callr</span> 3.7.0  <span style="color: #a3a3a3;">(115ms)</span>                                                
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">covr</span> 3.5.1  <span style="color: #a3a3a3;">(123ms)</span>                                                 
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">bslib</span> 0.3.1  <span style="color: #a3a3a3;">(160ms)</span>                                                
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">curl</span> 4.3.2  <span style="color: #a3a3a3;">(156ms)</span>                                                 
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">desc</span> 1.4.1  <span style="color: #a3a3a3;">(77ms)</span>                                                  
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">diffobj</span> 0.3.5  <span style="color: #a3a3a3;">(62ms)</span>                                               
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">digest</span> 0.6.29  <span style="color: #a3a3a3;">(40ms)</span>                                               
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">evaluate</span> 0.15  <span style="color: #a3a3a3;">(28ms)</span>                                               
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">fastmap</span> 1.1.0  <span style="color: #a3a3a3;">(14ms)</span>                                               
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">fs</span> 1.5.2  <span style="color: #a3a3a3;">(18ms)</span>                                                    
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">highr</span> 0.9  <span style="color: #a3a3a3;">(12ms)</span>                                                   
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">htmltools</span> 0.5.2  <span style="color: #a3a3a3;">(14ms)</span>                                             
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">htmlwidgets</span> 1.5.4  <span style="color: #a3a3a3;">(20ms)</span>                                           
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">httr</span> 1.4.3  <span style="color: #a3a3a3;">(18ms)</span>                                                  
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">jquerylib</span> 0.1.4  <span style="color: #a3a3a3;">(17ms)</span>                                             
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">jsonlite</span> 1.8.0  <span style="color: #a3a3a3;">(22ms)</span>                                              
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">knitr</span> 1.39  <span style="color: #a3a3a3;">(42ms)</span>                                                  
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">lazyeval</span> 0.2.2  <span style="color: #a3a3a3;">(14ms)</span>                                              
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">mime</span> 0.12  <span style="color: #a3a3a3;">(12ms)</span>                                                   
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">openssl</span> 2.0.2  <span style="color: #a3a3a3;">(33ms)</span>                                               
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">pkgload</span> 1.2.4  <span style="color: #a3a3a3;">(11ms)</span>                                               
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">praise</span> 1.0.0  <span style="color: #a3a3a3;">(10ms)</span>                                                
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">processx</span> 3.6.1  <span style="color: #a3a3a3;">(13ms)</span>                                              
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">ps</span> 1.7.1  <span style="color: #a3a3a3;">(13ms)</span>                                                    
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">rappdirs</span> 0.3.3  <span style="color: #a3a3a3;">(13ms)</span>                                              
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">rematch2</span> 2.1.2  <span style="color: #a3a3a3;">(13ms)</span>                                              
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">rex</span> 1.2.1  <span style="color: #a3a3a3;">(13ms)</span>                                                   
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">rmarkdown</span> 2.14  <span style="color: #a3a3a3;">(64ms)</span>                                              
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">rprojroot</span> 2.0.3  <span style="color: #a3a3a3;">(23ms)</span>                                             
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">rstudioapi</span> 0.13  <span style="color: #a3a3a3;">(15ms)</span>                                             
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">sass</span> 0.4.1  <span style="color: #a3a3a3;">(38ms)</span>                                                  
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">stringi</span> 1.7.6  <span style="color: #a3a3a3;">(149ms)</span>                                              
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">stringr</span> 1.4.0  <span style="color: #a3a3a3;">(70ms)</span>                                               
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">sys</span> 3.4  <span style="color: #a3a3a3;">(84ms)</span>                                                     
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">tinytex</span> 0.40  <span style="color: #a3a3a3;">(55ms)</span>                                                
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">testthat</span> 3.1.4  <span style="color: #a3a3a3;">(160ms)</span>                                             
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">waldo</span> 0.4.0  <span style="color: #a3a3a3;">(44ms)</span>                                                 
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">withr</span> 2.5.0  <span style="color: #a3a3a3;">(43ms)</span>                                                 
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">xfun</span> 0.31  <span style="color: #a3a3a3;">(69ms)</span>                                                   
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">yaml</span> 2.3.5  <span style="color: #a3a3a3;">(65ms)</span>                                                  
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">DT</span> 0.23  <span style="color: #a3a3a3;">(74ms)</span>                                                     
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">cachem</span> 1.0.6  <span style="color: #a3a3a3;">(15ms)</span>                                                
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">Rcpp</span> 1.0.8.3  <span style="color: #a3a3a3;">(111ms)</span>                                               
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">commonmark</span> 1.8.0  <span style="color: #a3a3a3;">(44ms)</span>                                            
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">crosstalk</span> 1.2.0  <span style="color: #a3a3a3;">(39ms)</span>                                             
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">fontawesome</span> 0.2.2  <span style="color: #a3a3a3;">(64ms)</span>                                           
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">formattable</span> 0.2.1  <span style="color: #a3a3a3;">(43ms)</span>                                           
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">httpuv</span> 1.6.5  <span style="color: #a3a3a3;">(49ms)</span>                                                
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">later</span> 1.3.0  <span style="color: #a3a3a3;">(45ms)</span>                                                 
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">promises</span> 1.2.0.1  <span style="color: #a3a3a3;">(40ms)</span>                                            
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">sourcetools</span> 0.1.7  <span style="color: #a3a3a3;">(19ms)</span>                                           
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">shiny</span> 1.7.1  <span style="color: #a3a3a3;">(81ms)</span>                                                 
#> <span style="color: #859900;">âœ”</span> Installed <span style="color: #268BD2;">xtable</span> 1.8-4  <span style="color: #a3a3a3;">(75ms)</span>                                                
#> <span style="color: #859900;">âœ”</span> 2 pkgs + 68 deps: kept 12, added 57, dld 57 (60.28 MB) <span style="color: #b8b8b8;">[17s]</span>                  
</pre></div>
}}


The syntax is

\if{html}{\out{<div class="sourceCode">}}\preformatted{<packagename>=?ignore
}\if{html}{\out{</div>}}

Note that you can only ignore \emph{optional} dependencies, i.e. packages in
\code{Suggests} and \code{Enhances}.
}
}

\section{Others}{
\subsection{How can I use pak with renv?}{

You cannot currently, but keep on eye on this issue:
https://github.com/r-lib/pak/issues/343
}
}

