% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{Get started with pak}
\alias{Get started with pak}
\title{Simplified manual. Start here!}
\description{
You don't need to read long manual pages for a simple task.
This manual page collects the most common pak use cases.

```{r child = "man/chunks/tldr.Rmd"}
```
}
