% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R, R/package.R
\name{lib_status}
\alias{lib_status}
\alias{pkg_list}
\title{Status of packages in a library}
\usage{
lib_status(lib = .libPaths()[1])

pkg_list(lib = .libPaths()[1])
}
\arguments{
\item{lib}{Path to library.}
}
\value{
Data frame the contains data about the packages
  installed in the library.
  \eval{include_docs("pkgdepends", "docs/lib-status-return.rds")}
}
\description{
Status of packages in a library
}
\section{Examples}{

```{asciicast lib-status-2}
lib_status(.Library)
```
}

\seealso{
Other package functions: 
\code{\link{pak}()},
\code{\link{pkg_deps_tree}()},
\code{\link{pkg_deps}()},
\code{\link{pkg_download}()},
\code{\link{pkg_install}()},
\code{\link{pkg_load}()},
\code{\link{pkg_remove}()},
\code{\link{pkg_status}()},
\code{\link{pkg_sysreqs}()}
}
\concept{library functions}
\concept{package functions}
