% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{lib_status}
\alias{lib_status}
\title{Status of packages in a library}
\usage{
lib_status(lib = .libPaths()[1])
}
\arguments{
\item{lib}{Path to library.}
}
\value{
Data frame (tibble) the contains data about the packages
installed in the library.
}
\description{
Status of packages in a library
}
\concept{library functions}
