% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R, R/package.R
\name{lib_status}
\alias{lib_status}
\alias{pkg_list}
\title{Status of packages in a library}
\usage{
lib_status(lib = .libPaths()[1])

pkg_list(lib = .libPaths()[1])
}
\arguments{
\item{lib}{Path to library.}
}
\value{
Data frame the contains data about the packages
installed in the library.
}
\description{
Status of packages in a library
}
\seealso{
Other package functions: 
\code{\link{pak}()},
\code{\link{pkg_deps_tree}()},
\code{\link{pkg_deps}()},
\code{\link{pkg_download}()},
\code{\link{pkg_install}()},
\code{\link{pkg_remove}()},
\code{\link{pkg_status}()}
}
\concept{library functions}
\concept{package functions}
