% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{local_pkg_install}
\alias{local_pkg_install}
\title{Install a local development package}
\usage{
local_pkg_install(path = ".", lib = .libPaths()[[1L]], num_workers = 1L)
}
\arguments{
\item{path}{to the local package}

\item{lib}{library to install packages into}

\item{num_workers}{Number of parallel workers to use}
}
\description{
Install a local development package
}
