% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-requirements.R
\name{local_system_requirements}
\alias{local_system_requirements}
\title{Query the system requirements for a dev package (and its dependencies)}
\usage{
local_system_requirements(root = ".", os, os_release)
}
\arguments{
\item{root}{Path to the package tree.}

\item{os, os_release}{The operating system and operating system release version, see
\url{https://github.com/rstudio/r-system-requirements#operating-systems} for the
list of supported operating systems.}
}
\value{
A character vector of commands needed to install the system requirements for the package.
}
\description{
Returns a character vector of commands to run that will install system
requirements for the queried operating system.
}
