% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{pak-config}
\alias{pak-config}
\alias{pak_config}
\title{pak configuration}
\description{
pak behavior can be finetuned with environment variables and
options (as in \code{\link[base:options]{base::options()}}).
}
\details{
Configuration entries (unless noted otherwise on this manual page)
have a corresponding environment variable, and a corresponding option.

The environment variable is always uppercase and uses underscores as the
word separator. It always has the \code{PKG_} prefix.

The option is typically lowercase, use it uses underscores as the word
separator, but it always has the \code{pkg.} prefix (notice the dot!).

Some examples:\tabular{lll}{
   Config entry name \tab Env var name \tab Option name \cr
   platforms \tab \code{PKG_PLATFORMS} \tab \code{pkg.platforms} \cr
   cran_mirror \tab \code{PKG_CRAN_MIRROR} \tab \code{pkg.cran_mirror} \cr
}

\subsection{pak configuration entries}{

\eval{pak:::doc_config()}
}
}
