% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{package}
\name{pak-package}
\alias{pak-package}
\alias{_PACKAGE}
\alias{.about}
\title{A fresh approach to package installation}
\description{
The goal of 'pak' is to make package installation faster and
    more reliable. In particular, it performs all HTTP operations in parallel,
    so metadata resolution and package downloads are fast. Metadata and package
    files are cached on the local disk as well. 'pak' has a dependency solver,
    so it finds version conflicts before performing the installation. This
    version of 'pak' supports CRAN, 'Bioconductor' and 'GitHub' packages as well.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pak.r-lib.org/}
  \item Report bugs at \url{https://github.com/r-lib/pak/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Jim Hester \email{jim.hester@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
