% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pak-cleanup.R
\name{pak_cleanup}
\alias{pak_cleanup}
\title{Clean up pak caches and/or the pak library}
\usage{
pak_cleanup(package_cache = TRUE, metadata_cache = TRUE,
  pak_lib = TRUE, force = FALSE)
}
\arguments{
\item{package_cache}{Whether to clean up the cache of package files.}

\item{metadata_cache}{Whether to clean up the cache of package meta
data.}

\item{pak_lib}{Whether to clean up the pak package library.}

\item{force}{Do not ask for confirmation. Note that to use this function
in non-interactive mode, you have to specify \code{force = FALSE}.}
}
\description{
Clean up pak caches and/or the pak library
}
\seealso{
Other pak housekeeping: \code{\link{pak_private_library}},
  \code{\link{pak_setup}}, \code{\link{pak_sitrep}}
}
\concept{pak housekeeping}
