% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pak_package_sources}
\alias{pak_package_sources}
\title{Package sources}
\description{
Learn how to tell pak which packages to install, and where those packages
can be found.
}
\seealso{
Other package functions: 
\code{\link{lib_status}()},
\code{\link{pak}()},
\code{\link{pkg_deps_tree}()},
\code{\link{pkg_deps}()},
\code{\link{pkg_download}()},
\code{\link{pkg_install}()},
\code{\link{pkg_remove}()},
\code{\link{pkg_status}()}
}
\concept{package functions}
\eval{pak:::include_docs("pkgdepends", "docs/pkg-refs.rds")}
