% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{Package sources}
\alias{Package sources}
\title{Install packages from CRAN, Bioconductor, GitHub, URLs, etc.}
\description{
Install packages from CRAN, Bioconductor, GitHub, URLs, etc.
Learn how to tell pak which packages to install, and where those packages
can be found.

If you want a quick overview of package sources, see
"\link{Get started with pak}".
}
\details{
\eval{pak:::include_docs("pkgdepends", "docs/pkg-refs.rds", top = FALSE)}
}
