% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pak-install.R
\name{pak_sitrep}
\alias{pak_sitrep}
\alias{pak_sitrep_data}
\title{pak SITuation REPort}
\usage{
pak_sitrep()
}
\description{
It prints
* pak version,
* the current library path,
* location of the private library,
* whether the pak private library exists,
* whether the pak private library is functional.
}
\section{Examples}{

```{asciicast pak-sitrep}
pak_sitrep()
```
}

\seealso{
Other pak housekeeping: 
\code{\link{pak_cleanup}()}
}
\concept{pak housekeeping}
