% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pak-install.R
\name{pak_sitrep}
\alias{pak_sitrep}
\title{pak SITuation REPort}
\usage{
pak_sitrep()
}
\description{
It prints
\itemize{
\item pak version,
\item the current library path,
\item location of the private library,
\item whether the pak private library exists,
\item whether the pak private library is functional.
}
}
\seealso{
Other pak housekeeping: \code{\link{pak_cleanup}},
  \code{\link{pak_private_library}},
  \code{\link{pak_setup}}
}
\concept{pak housekeeping}
