% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pak-update.R
\name{pak_update}
\alias{pak_update}
\title{Update pak itself}
\usage{
pak_update(force = FALSE, stream = c("auto", "stable", "rc", "devel"))
}
\arguments{
\item{force}{Whether to force an update, even if no newer version is
available.}

\item{stream}{Whether to update to the
  * `"stable"`,
  * `"rc"` (release candidate) or
  * `"devel"` (development) version.
  * `"auto"` updates to the same stream as the current one.

  Often there is no release candidate version, then `"rc"` also
  installs the stable version.}
}
\value{
Nothing.
}
\description{
Use this function to update the released or development version of pak.
}
