% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_deps}
\alias{pkg_deps}
\title{Look up the dependencies of a package}
\usage{
pkg_deps(pkg, upgrade = TRUE, dependencies = NA)
}
\arguments{
\item{pkg}{Package names or package references. E.g.
  - `ggplot2`: package from CRAN, Bioconductor or a CRAN-like repository
    in general,
  - `tidyverse/ggplot2`: package from GitHub,
  - `tidyverse/ggplot2@v3.4.0`: package from GitHub tag or branch,
  - `https://examples.com/.../ggplot2_3.3.6.tar.gz`: package from URL,
  - `.`: package in the current working directory.

  See "[Package sources]" for more details.}

\item{upgrade}{Whether to use the most recent available package
versions.}

\item{dependencies}{What kinds of dependencies to install. Most commonly
one of the following values:
- `NA`: only required (hard) dependencies,
- `TRUE`: required dependencies plus optional and development
  dependencies,
- `FALSE`: do not install any dependencies. (You might end up with a
  non-working package, and/or the installation might fail.)
See [Package dependency types] for other possible values and more
information about package dependencies.}
}
\value{
A data frame with the dependency data, it includes `pkg`
  as well. It has the following columns.
  \eval{include_docs("pkgdepends", "docs/resolution-result.rds")}
}
\description{
Look up the dependencies of a package
}
\section{Examples}{

```{asciicast pkg-deps}
pkg_deps("dplyr")
```

For a package on GitHub:
```{asciicast pkg-deps-gh}
pkg_deps("r-lib/callr")
```
}

\seealso{
Other package functions: 
\code{\link{lib_status}()},
\code{\link{pak}()},
\code{\link{pkg_deps_tree}()},
\code{\link{pkg_download}()},
\code{\link{pkg_install}()},
\code{\link{pkg_load}()},
\code{\link{pkg_remove}()},
\code{\link{pkg_status}()},
\code{\link{pkg_sysreqs}()}
}
\concept{package functions}
