% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_deps}
\alias{pkg_deps}
\title{Look up the dependencies of a package}
\usage{
pkg_deps(pkg, upgrade = TRUE, dependencies = NA)
}
\arguments{
\item{pkg}{Package name or remote package specification to resolve.}

\item{upgrade}{Whether to use the most recent available package
versions.}

\item{dependencies}{Dependency types. See
\code{\link[pkgdepends:as_pkg_dependencies]{pkgdepends::as_pkg_dependencies()}} for possible values.}
}
\value{
A data frame (tibble).
}
\description{
Look up the dependencies of a package
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pkg_deps("curl")
pkg_deps("r-lib/fs")
\dontshow{\}) # examplesIf}
}
\seealso{
Other package functions: 
\code{\link{lib_status}()},
\code{\link{pak_package_sources}},
\code{\link{pak}()},
\code{\link{pkg_deps_tree}()},
\code{\link{pkg_download}()},
\code{\link{pkg_install}()},
\code{\link{pkg_remove}()},
\code{\link{pkg_status}()}
}
\concept{package functions}
