% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{pkg_history}
\alias{pkg_history}
\title{Query the history of a CRAN package}
\usage{
pkg_history(pkg)
}
\arguments{
\item{pkg}{Package name.}
}
\value{
A data frame, with one row per package version. The columns are
  the entries of the `DESCRIPTION` files in the released package
  versions.
}
\description{
Query the history of a CRAN package
}
\section{Examples}{

```{asciicast pkg-history, R.options = list(width = 72)}
pkg_history("ggplot2")
```
}

