% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{pkg_history}
\alias{pkg_history}
\title{Query the history of a package}
\usage{
pkg_history(pkg)
}
\arguments{
\item{pkg}{Package name.}
}
\value{
A data frame (tibble), with one row per package version.
}
\description{
Query the history of a package
}
