% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_install}
\alias{pkg_install}
\title{Install a package}
\usage{
pkg_install(pkg, lib = .libPaths()[[1L]], upgrade = FALSE,
  num_workers = 1L)
}
\arguments{
\item{pkg}{A package specification. See \link[pkgdepends:remotes]{pkgdepends::remotes} for details.}

\item{lib}{library to install packages into}

\item{upgrade}{whether to upgrade already installed packages to the
latest available version}

\item{num_workers}{Number of parallel workers to use}
}
\description{
Install a package and it's dependencies.
}
