% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_remove}
\alias{pkg_remove}
\title{Remove installed packages}
\usage{
pkg_remove(pkg, lib = .libPaths()[[1L]])
}
\arguments{
\item{pkg}{A character vector of packages to remove.}

\item{lib}{library to remove packages from.}
}
\value{
Nothing.
}
\description{
Remove installed packages
}
\seealso{
Other package functions: 
\code{\link{lib_status}()},
\code{\link{pak}()},
\code{\link{pkg_deps_tree}()},
\code{\link{pkg_deps}()},
\code{\link{pkg_download}()},
\code{\link{pkg_install}()},
\code{\link{pkg_status}()},
\code{\link{pkg_sysreqs}()}
}
\concept{package functions}
