% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{pkg_search}
\alias{pkg_search}
\title{Search CRAN packages}
\usage{
pkg_search(query, ...)
}
\arguments{
\item{query}{Search query string.}

\item{...}{Additional arguments passed to [pkgsearch::pkg_search()]}
}
\value{
A data frame, that is also a `pak_search_result` object
with a custom print method. To see the underlying table, you
can use `[]` to drop the extra classes. See examples below.
}
\description{
Search the indexed database of current CRAN packages. It uses the
pkgsearch package. See that package for more details and also
[pkgsearch::pkg_search()] for pagination, more advanced searching,
etc.
}
\section{Examples}{

Simple search
```{asciicast pkg-search, R.options = list(width = 72)}
pkg_search("survival")
```

See the underlying data frame
```{asciicast pkg-search-2, R.options = list(width = 72)}
psro <- pkg_search("ropensci")
psro[]
```
}

