% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{pkg_search}
\alias{pkg_search}
\title{Search CRAN packages}
\usage{
pkg_search(query, ...)
}
\arguments{
\item{query}{Search query string.}

\item{...}{Additional arguments passed to \code{\link[pkgsearch:pkg_search]{pkgsearch::pkg_search()}}}
}
\value{
A data frame (tibble), that is also a \code{pak_search_result}
object with a custom print method. To see the underlying table, you
can use \verb{[]} to drop the extra classes. See examples below.
}
\description{
Search the indexed database of current CRAN packages. It uses the
pkgsearch package. See that package for more details and also
\code{\link[pkgsearch:pkg_search]{pkgsearch::pkg_search()}} for pagination, more advanced searching,
etc.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Simple search
pkg_search("survival")

# See the underlying tibble
psro <- pkg_search("ropensci")
psro[]
\dontshow{\}) # examplesIf}
}
