% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_status}
\alias{pkg_status}
\title{Display installed locations of a package}
\usage{
pkg_status(pkg, lib = .libPaths())
}
\arguments{
\item{pkg}{Name of one or more installed packages to display status for.}

\item{lib}{One or more library paths to lookup packages status in.
By default all libraries are used.}
}
\value{
Data frame with data about installations of `pkg`.
  \eval{include_docs("pkgdepends", "docs/lib-status-return.rds")}
}
\description{
Display installed locations of a package
}
\section{Examples}{

```{asciicast pkg-status}
pkg_status("MASS")
```
}

\seealso{
Other package functions: 
\code{\link{lib_status}()},
\code{\link{pak}()},
\code{\link{pkg_deps_tree}()},
\code{\link{pkg_deps}()},
\code{\link{pkg_download}()},
\code{\link{pkg_install}()},
\code{\link{pkg_load}()},
\code{\link{pkg_remove}()},
\code{\link{pkg_sysreqs}()}
}
\concept{package functions}
