% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_status}
\alias{pkg_status}
\title{Display installed locations of a package}
\usage{
pkg_status(pkg, lib = .libPaths())
}
\arguments{
\item{pkg}{Name of one or more installed packages to display status for.}

\item{lib}{One or more library paths to lookup packages status in.}
}
\value{
Data frame with data about installations of \code{pkg}.
Columns include: \code{library}, \code{package}, \code{title}, \code{version}.
}
\description{
Display installed locations of a package
}
\examples{
\dontrun{
pkg_status("MASS")
}
}
\seealso{
Other package functions: 
\code{\link{lib_status}()},
\code{\link{pak}()},
\code{\link{pkg_deps_tree}()},
\code{\link{pkg_deps}()},
\code{\link{pkg_download}()},
\code{\link{pkg_install}()},
\code{\link{pkg_remove}()}
}
\concept{package functions}
