% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{ppm_has_binaries}
\alias{ppm_has_binaries}
\title{Does PPM build binary packages for the current platform?}
\usage{
ppm_has_binaries()
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Does PPM build binary packages for the current platform?
}
\examples{
\dontshow{if (!pkgcache:::is_rcmd_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
system_r_platform()
ppm_has_binaries()
\dontshow{\}) # examplesIf}
}
\seealso{
The 'pkgcache and Posit Package Manager on Linux'
article at \url{https://r-lib.github.io/pkgcache/dev}.

Other PPM functions: 
\code{\link{ppm_platforms}()},
\code{\link{ppm_r_versions}()},
\code{\link{ppm_repo_url}()},
\code{\link{ppm_snapshots}()}
}
\concept{PPM functions}
