% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_add}
\alias{repo_add}
\alias{repo_resolve}
\title{Add a new CRAN-like repository}
\usage{
repo_add(..., .list = NULL)

repo_resolve(spec)
}
\arguments{
\item{...}{Repository specifications, possibly named character vectors.
See details below.}

\item{.list}{List or character vector of repository specifications.
This argument is easier to use programmatically than `...`. See
details below.}

\item{spec}{Repository specification, a possibly named character
scalar.}
}
\value{
`repo_resolve()` returns a named character scalar, the URL
of the repository.
}
\description{
Add a new repository to the list of repositories that pak uses to
look for packages.
}
\details{
`repo_add()` adds new repositories. It resolves the specified
repositories using `repo_resolve()` and then modifies the `repos`
global option.

`repo_add()` only has an effect in the current R session. If you
want to keep your configuration between R sessions, then set the
`repos` option to the desired value in your user or project `.Rprofile`
file.


# Repository specifications

The format of a repository specification is a named or unnamed
character scalar. If the name is missing, pak adds a name
automatically. The repository named `CRAN` is the main CRAN repository,
but otherwise names are informational.

Currently supported repository specifications:
- URL pointing to the root of the CRAN-like repository. Example:
  ```
  https://cloud.r-project.org
  ```
- `RSPM@<date>`, RSPM (RStudio Package Manager) snapshot, at the
  specified date.
- `RSPM@<package>-<version>` RSPM snapshot, for the day after the
  release of `<version>` of `<package>`.
- `RSPM@R-<version>` RSPM snapshot, for the day after R `<version>`
  was released.
- `MRAN@<date>`, MRAN (Microsoft R Application Network) snapshot, at
  the specified date.
- `MRAN@<package>-<version>` MRAN snapshot, for the
  day after the release of `<version>` of `<package>`.
- `MRAN@R-<version>` MRAN snapshot, for the day
  after R `<version>` was released.


Notes:
* See more about RSPM at `https://packagemanager.rstudio.com/client/#/`.
* See more about MRAN snapshots at
  <https://mran.microsoft.com/timemachine>.
* All dates (or times) can be specified in the ISO 8601 format.
* If RSPM does not have a snapshot available for a date, the next
  available date is used.
* Dates that are before the first, or after the last RSPM snapshot
  will trigger an error.
* Dates before the first, or after the last MRAN snapshot will trigger
  an error.
* Unknown R or package versions will trigger an error.
}
\section{Exaples}{

```{asciicast repo-add}
repo_add(RSPMdplyr100 = "RSPM@dplyr-1.0.0")
repo_get()
```

```{asciicast repo-add-2}
repo_resolve("MRAN@2020-01-21")
```
```{asciicast repo-add-3}
repo_resolve("RSPM@2020-01-21")
```
```{asciicast repo-add-4}
repo_resolve("MRAN@dplyr-1.0.0")
```
```{asciicast repo-add-5}
repo_resolve("RSPM@dplyr-1.0.0")
```
```{asciicast repo-add-6}
repo_resolve("MRAN@R-4.0.0")
```
```{asciicast repo-add-7}
repo_resolve("RSPM@R-4.0.0")
```
}

\seealso{
Other repository functions: 
\code{\link{repo_get}()},
\code{\link{repo_status}()}
}
\concept{repository functions}
