% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{repo_auth_key_get}
\alias{repo_auth_key_get}
\alias{repo_auth_key_set}
\title{Query or set repository password in the system credential store}
\usage{
repo_auth_key_get(url, username = NULL)

repo_auth_key_set(url, password, username = NULL)
}
\arguments{
\item{url}{Repository URL. It may contain a username, in which case
\code{username} may be \code{NULL}.}

\item{username}{User name, if it is not included in \code{url}.}

\item{password}{Password to set.}
}
\value{
\code{repo_auth_key_get()} returns a single string, the repository
password.
}
\description{
Use pak's keyring functions to query or set a repository password in
the system credential store.
}
\details{
\code{repo_auth_key_get()} retrieves a password from the default keyring. It
errors if it cannot find the credentials for \code{url}.

\code{repo_auth_key_set()} adds or updates a password in the system
credential store.
}
\seealso{
Other authenticated repositories: 
\code{\link{Authenticated repositories}},
\code{\link{repo_auth}()}
}
\concept{authenticated repositories}
