% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_status}
\alias{repo_status}
\alias{repo_ping}
\title{Show the status of CRAN-like repositories}
\usage{
repo_status(
  platforms = NULL,
  r_version = getRversion(),
  bioc = TRUE,
  cran_mirror = NULL
)

repo_ping(
  platforms = NULL,
  r_version = getRversion(),
  bioc = TRUE,
  cran_mirror = NULL
)
}
\arguments{
\item{platforms}{Platforms to use, default is the current platform,
plus source packages.}

\item{r_version}{R version(s) to use, the default is the current
R version, via [getRversion()].}

\item{bioc}{Whether to add the Bioconductor repositories. If you
already configured them via `options(repos)`, then you can
set this to `FALSE`.}

\item{cran_mirror}{The CRAN mirror to use.}
}
\value{
A data frame that has a row for every repository, on every
queried platform and R version. It has these columns:
* `name`: the name of the repository. This comes from the names
  of the configured repositories in `options("repos")`, or
  added by pkgcache. It is typically `CRAN` for CRAN, and the
  current Bioconductor repositories are `BioCsoft`, `BioCann`,
  `BioCexp`, `BioCworkflows`.
* `url`: base URL of the repository.
* `bioc_version`: Bioconductor version, or `NA` for
  non-Bioconductor repositories.
* `platform`: platform, possible values are `source`, `macos` and
  `windows` currently.
* `path`: the path to the packages within the base URL, for a
  given platform and R version.
* `r_version`: R version, one of the specified R versions.
* `ok`: Logical flag, whether the repository contains a metadata
  file for the given platform and R version.
* `ping`: HTTP response time of the repository in seconds. If
  the `ok` column is `FALSE`, then this columns in `NA`.
* `error`: the error object if the HTTP query failed for this
  repository, platform and R version.
}
\description{
It checks the status of the configured or supplied repositories.
}
\details{
`repo_ping()` is similar to `repo_status()` but also prints a short
summary of the data, and it returns its result invisibly.
}
\section{Examples}{

```{asciicast repo-status}
repo_status()
```

```{asciicast repo-status-2}
repo_status(
  platforms = c("windows", "macos"),
  r_version = c("4.0", "4.1")
)
```

```{asciicast repo-status-3, asciicast_cols = 77, asciicast_width = 77, R.options = list(cli.width = 77, width = 77)}
repo_ping()
```
}

\seealso{
Other repository functions: 
\code{\link{repo_add}()},
\code{\link{repo_get}()}
}
\concept{repository functions}
