% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{System requirements}
\alias{System requirements}
\title{System requirements}
\description{
pak takes care of your system requirements.
}
\section{Supported platforms}{
TODO
}

\section{About other OSes}{
\subsection{Windows}{

TODO
}

\subsection{macOS}{

TODO
}
}

