% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysreqsdocs.R
\name{sysreqs_check_installed}
\alias{sysreqs_check_installed}
\title{Check if installed packages have all their system requirements}
\usage{
sysreqs_check_installed(packages = NULL, library = .libPaths()[1])
}
\arguments{
\item{packages}{If not \code{NULL}, then only these packages are checked.
If a package in \code{packages} is not installed, then
pak throws a warning.}

\item{library}{Library or libraries to check.}
}
\value{
Data frame with a custom print and format method, and a
\code{pkg_sysreqs_check_result} class. Its columns are:
\itemize{
\item \code{system_package}: string, name of the required system package.
\item \code{installed}: logical, whether the system package is correctly
installed.
\item \code{packages}: list column of character vectors. The names of the
installed R packages that need this system package.
}
}
\description{
Check if installed packages have all their system requirements
}
\details{
This function uses the \code{sysreqs_platform} configuration option,
see \eval{man_config_link("Configuration")}. Set this if
pak does not detect your platform correctly.
}
\examples{
\dontshow{if (Sys.getenv("IN_PKGDOWN") == "true" && Sys.info()[["sysname"]] == "Linux") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# This only works on supported platforms
sysreqs_check_installed()
\dontshow{\}) # examplesIf}
}
\seealso{
Other system requirements functions: 
\code{\link{pkg_sysreqs}()},
\code{\link{sysreqs_db_list}()},
\code{\link{sysreqs_db_match}()},
\code{\link{sysreqs_db_update}()},
\code{\link{sysreqs_is_supported}()},
\code{\link{sysreqs_list_system_packages}()},
\code{\link{sysreqs_platforms}()}
}
\concept{system requirements functions}
